Fixed collision bug caused by float precision

+++ MinecraftDemo/src/Core/Chunk/Chunk.h
@@ -45,7 +45,7 @@ public:
 	static void deleteChunkData(Cube*** chunkData, int height, int width);
 	Cube*** getBlockMatrix();
 	void setPosition(glm::vec3& position);
-	Cube* getCubeByCoords(glm::vec3 coords);
+	Cube* getCubeByCoords(glm::f64vec3 coords);
 	Cube* findNeighbourBlock(Cube::FaceSide neighbourSide, int height, int width, int depth);
 	Cube* findNeighbourBlock(Cube::FaceSide neighbourSide, Cube* cube);
 	ChunkRenderData& getChunkRenderData();
