+++ MinecraftDemo/src/Core/Services/MovementSystem/MovementModes/StandardMovementMode.cpp
@@ -37,14 +37,14 @@ void StandardMovementMode::stepPhysicalMovement() {
 	mc->m_previousPosition = mc->m_currentPosition;
 	mc->m_previousVelocity = mc->m_currentVelocity;
 
-	glm::vec3 targetVelNorm(0.0f, 0.0f, 0.0f);
-	if (glm::length(mc->m_targetVelocity) > 0.0f) {
+	glm::f64vec3 targetVelNorm(0.0, 0.0, 0.0);
+	if (glm::length(mc->m_targetVelocity) > 0.0) {
 		targetVelNorm = glm::normalize(mc->m_targetVelocity);
 	}
 
-	float gravity = GRAVITY * mc->m_gravityInfluence;
+	double gravity = GRAVITY * mc->m_gravityInfluence;
 	mc->m_velocity.y = mc->m_velocity.y + (-gravity * mc->m_gravityInfluence * FIXED_DELTA);
-	float lerpFactor = mc->m_isGrounded ? GROUND_VELOCITY_LERP_FACTOR * mc->m_groundDampeningInfluence : AIR_VELOCITY_LERP_FACTOR * mc->m_airDampeningInfluence;
+	double lerpFactor = mc->m_isGrounded ? GROUND_VELOCITY_LERP_FACTOR * mc->m_groundDampeningInfluence : AIR_VELOCITY_LERP_FACTOR * mc->m_airDampeningInfluence;
 	//mc->m_velocity.z = cubicInterpolation();
 	mc->m_velocity.z = mc->m_velocity.z + ((mc->m_targetVelocity.z - mc->m_velocity.z) * lerpFactor);
 	mc->m_velocity.x = mc->m_velocity.x + ((mc->m_targetVelocity.x - mc->m_velocity.x) * lerpFactor);
@@ -72,7 +72,7 @@ void StandardMovementMode::stepPhysicalMovement() {
 	}
 	mc->m_velocity.x = mc->m_velocity.x + velocityDecrementX;
 	*/
-	glm::vec3 targetPosition = mc->m_currentPosition;
+	glm::f64vec3 targetPosition = mc->m_currentPosition;
 	collider.buildAABBCollider(mc->m_currentPosition, mc->m_colliderHalfWidth, mc->m_colliderHalfHeight);
 
 	targetPosition.y += mc->m_velocity.y * FIXED_DELTA;
@@ -80,9 +80,9 @@ void StandardMovementMode::stepPhysicalMovement() {
 
 	//BOOST_LOG_TRIVIAL(info) << "aabb: " << collider.bottomRightBack.x << " " << collider.bottomRightBack.y << " " << collider.bottomRightBack.z;
 
-	glm::vec3 targetDistance = targetPosition;
+	glm::f64vec3 targetDistance = targetPosition;
 
-	glm::vec3 colliderVertexPosition = collider.bottomLeftFront;
+	glm::f64vec3 colliderVertexPosition = collider.bottomLeftFront;
 	bool vertical = true;
 	bool horizontal = true;
 	bool depth = true;
@@ -93,20 +93,20 @@ void StandardMovementMode::stepPhysicalMovement() {
 	bool leftCollision = false;
 	bool rightCollision = false;
 	Cube::CubeId cubeId;
-	float nearestCubes = 0.0f;
+	double nearestCubes = 0.0;
 	// Bottom side
-	if (mc->m_velocity.y < 0.0f) {
+	if (mc->m_velocity.y < 0.0) {
 		targetDistance.y -= mc->m_colliderHalfHeight;
 		//BOOST_LOG_TRIVIAL(info) << "POS2: " << targetPosition.y;
 		//BOOST_LOG_TRIVIAL(info) << "DISTANCE: " << glm::abs(targetDistance.y - colliderVertexPosition.y) << " " << mc->m_velocity.y << " " << targetDistance.y;
 		do {
 			nearestCubes = glm::floor(colliderVertexPosition.y);
-			if (glm::abs(targetDistance.y - colliderVertexPosition.y) < 1.0f) {
+			if (glm::abs(targetDistance.y - colliderVertexPosition.y) < 1.0) {
 				colliderVertexPosition.y = targetDistance.y;
 				depth = false;
 			}
 			else {
-				colliderVertexPosition.y += (glm::sign(mc->m_velocity.y) * 1.0f);
+				colliderVertexPosition.y += (glm::sign(mc->m_velocity.y) * 1.0);
 			}
 			//BOOST_LOG_TRIVIAL(info) << "sub: " << colliderVertexPosition.y;
 			if (colliderVertexPosition.y < nearestCubes) {
@@ -149,15 +149,15 @@ void StandardMovementMode::stepPhysicalMovement() {
 							vertical = false;
 						}
 						*/
-						colliderVertexPosition.x += 1.0f;
+						colliderVertexPosition.x += 1.0;
 					}
-					colliderVertexPosition.z -= 1.0f;
+					colliderVertexPosition.z -= 1.0;
 				}
 			}
 		} while (depth);
 	}
 	//Top side
-	if (mc->m_velocity.y > 0.0f) {
+	if (mc->m_velocity.y > 0.0) {
 		targetDistance.y += mc->m_colliderHalfHeight;
 		colliderVertexPosition = collider.topLeftBack;
 		vertical = true;
@@ -165,12 +165,12 @@ void StandardMovementMode::stepPhysicalMovement() {
 		depth = true;
 		do {
 			nearestCubes = glm::ceil(colliderVertexPosition.y);
-			if (glm::abs(targetDistance.y - colliderVertexPosition.y) < 1.0f) {
+			if (glm::abs(targetDistance.y - colliderVertexPosition.y) < 1.0) {
 				colliderVertexPosition.y = targetDistance.y;
 				depth = false;
 			}
 			else {
-				colliderVertexPosition.y += (glm::sign(mc->m_velocity.y) * 1.0f);
+				colliderVertexPosition.y += (glm::sign(mc->m_velocity.y) * 1.0);
 			}
 			if (colliderVertexPosition.y > nearestCubes) {
 				colliderVertexPosition.z = collider.topLeftBack.z;
@@ -195,9 +195,9 @@ void StandardMovementMode::stepPhysicalMovement() {
 							vertical = false;
 							depth = false;
 						}
-						colliderVertexPosition.x += 1.0f;
+						colliderVertexPosition.x += 1.0;
 					}
-					colliderVertexPosition.z += 1.0f;
+					colliderVertexPosition.z += 1.0;
 				}
 			}
 		} while (depth);
@@ -210,7 +210,7 @@ void StandardMovementMode::stepPhysicalMovement() {
 	//BOOST_LOG_TRIVIAL(info) << "gz: " << go->m_position.x << " " << go->m_position.y << " " << go->m_position.z;
 	targetDistance = targetPosition;
 	//Back side
-	if (mc->m_velocity.z < 0.0f) {
+	if (mc->m_velocity.z < 0.0) {
 		targetDistance.z -= mc->m_colliderHalfWidth;
 		colliderVertexPosition = collider.topRightBack;
 		vertical = true;
@@ -218,12 +218,12 @@ void StandardMovementMode::stepPhysicalMovement() {
 		depth = true;
 		do {
 			nearestCubes = glm::floor(colliderVertexPosition.z);
-			if (glm::abs(targetDistance.z - colliderVertexPosition.z) < 1.0f) {
+			if (glm::abs(targetDistance.z - colliderVertexPosition.z) < 1.0) {
 				colliderVertexPosition.z = targetDistance.z;
 				depth = false;
 			}
 			else {
-				colliderVertexPosition.z += (glm::sign(mc->m_velocity.z) * 1.0f);
+				colliderVertexPosition.z += (glm::sign(mc->m_velocity.z) * 1.0);
 			}
 			if (colliderVertexPosition.z < nearestCubes) {
 				colliderVertexPosition.y = collider.topRightBack.y;
@@ -249,15 +249,15 @@ void StandardMovementMode::stepPhysicalMovement() {
 							vertical = false;
 							depth = false;
 						}
-						colliderVertexPosition.x -= 1.0f;
+						colliderVertexPosition.x -= 1.0;
 					}
-					colliderVertexPosition.y -= 1.0f;
+					colliderVertexPosition.y -= 1.0;
 				}
 			}
 		} while (depth);
 	}
 	//Front side
-	if (mc->m_velocity.z > 0.0f) {
+	if (mc->m_velocity.z > 0.0) {
 		targetDistance.z += mc->m_colliderHalfWidth;
 		colliderVertexPosition = collider.topLeftFront;
 		vertical = true;
@@ -265,12 +265,12 @@ void StandardMovementMode::stepPhysicalMovement() {
 		depth = true;
 		do {
 			nearestCubes = glm::ceil(colliderVertexPosition.z);
-			if (glm::abs(targetDistance.z - colliderVertexPosition.z) < 1.0f) {
+			if (glm::abs(targetDistance.z - colliderVertexPosition.z) < 1.0) {
 				colliderVertexPosition.z = targetDistance.z;
 				depth = false;
 			}
 			else {
-				colliderVertexPosition.z += (glm::sign(mc->m_velocity.z) * 1.0f);
+				colliderVertexPosition.z += (glm::sign(mc->m_velocity.z) * 1.0);
 			}
 			if (colliderVertexPosition.z > nearestCubes) {
 				colliderVertexPosition.y = collider.topLeftFront.y;
@@ -295,9 +295,9 @@ void StandardMovementMode::stepPhysicalMovement() {
 							vertical = false;
 							depth = false;
 						}
-						colliderVertexPosition.x += 1.0f;
+						colliderVertexPosition.x += 1.0;
 					}
-					colliderVertexPosition.y -= 1.0f;
+					colliderVertexPosition.y -= 1.0;
 				}
 			}
 		} while (depth);
@@ -309,7 +309,7 @@ void StandardMovementMode::stepPhysicalMovement() {
 	//BOOST_LOG_TRIVIAL(info) << "gx: " << go->m_position.x << " " << go->m_position.y << " " << go->m_position.z;
 	targetDistance = targetPosition;
 	//Left side
-	if (mc->m_velocity.x < 0.0f) {
+	if (mc->m_velocity.x < 0.0) {
 		targetDistance.x -= mc->m_colliderHalfWidth;
 		colliderVertexPosition = collider.topLeftBack;
 		vertical = true;
@@ -317,12 +317,12 @@ void StandardMovementMode::stepPhysicalMovement() {
 		depth = true;
 		do {
 			nearestCubes = glm::floor(colliderVertexPosition.x);
-			if (glm::abs(targetDistance.x - colliderVertexPosition.x) < 1.0f) {
+			if (glm::abs(targetDistance.x - colliderVertexPosition.x) < 1.0) {
 				colliderVertexPosition.x = targetDistance.x;
 				depth = false;
 			}
 			else {
-				colliderVertexPosition.x += (glm::sign(mc->m_velocity.x) * 1.0f);
+				colliderVertexPosition.x += (glm::sign(mc->m_velocity.x) * 1.0);
 			}
 			if (colliderVertexPosition.x < nearestCubes) {
 				colliderVertexPosition.y = collider.topLeftBack.y;
@@ -347,15 +347,15 @@ void StandardMovementMode::stepPhysicalMovement() {
 							vertical = false;
 							depth = false;
 						}
-						colliderVertexPosition.z += 1.0f;
+						colliderVertexPosition.z += 1.0;
 					}
-					colliderVertexPosition.y -= 1.0f;
+					colliderVertexPosition.y -= 1.0;
 				}
 			}
 		} while (depth);
 	}
 	//Right side
-	if (mc->m_velocity.x > 0.0f) {
+	if (mc->m_velocity.x > 0.0) {
 		targetDistance.x += mc->m_colliderHalfWidth;
 		colliderVertexPosition = collider.topRightFront;
 		vertical = true;
@@ -363,12 +363,12 @@ void StandardMovementMode::stepPhysicalMovement() {
 		depth = true;
 		do {
 			nearestCubes = glm::ceil(colliderVertexPosition.x);
-			if (glm::abs(targetDistance.x - colliderVertexPosition.x) < 1.0f) {
+			if (glm::abs(targetDistance.x - colliderVertexPosition.x) < 1.0) {
 				colliderVertexPosition.x = targetDistance.x;
 				depth = false;
 			}
 			else {
-				colliderVertexPosition.x += (glm::sign(mc->m_velocity.x) * 1.0f);
+				colliderVertexPosition.x += (glm::sign(mc->m_velocity.x) * 1.0);
 			}
 			if (colliderVertexPosition.x > nearestCubes) {
 				colliderVertexPosition.y = collider.topRightFront.y;
@@ -394,9 +394,9 @@ void StandardMovementMode::stepPhysicalMovement() {
 							vertical = false;
 							depth = false;
 						}
-						colliderVertexPosition.z -= 1.0f;
+						colliderVertexPosition.z -= 1.0;
 					}
-					colliderVertexPosition.y -= 1.0f;
+					colliderVertexPosition.y -= 1.0;
 				}
 			}
 		} while (depth);
@@ -405,34 +405,34 @@ void StandardMovementMode::stepPhysicalMovement() {
 
 
 	if (bottomCollision) {
-		mc->m_velocity.y = 0.0f;
+		mc->m_velocity.y = 0.0;
 		mc->m_isGrounded = true;
 	}
 	else {
 		mc->m_isGrounded = false;
 	}
 	if (topCollision) {
-		mc->m_velocity.y = 0.0f;
+		mc->m_velocity.y = 0.0;
 	}
 	if (leftCollision) {
-		mc->m_velocity.x = 0.0f;
+		mc->m_velocity.x = 0.0;
 	}
 	if (rightCollision) {
-		mc->m_velocity.x = 0.0f;
+		mc->m_velocity.x = 0.0;
 	}
 	if (frontCollision) {
-		mc->m_velocity.z = 0.0f;
+		mc->m_velocity.z = 0.0;
 	}
 	if (backCollision) {
-		mc->m_velocity.z = 0.0f;
+		mc->m_velocity.z = 0.0;
 	}
 
 	mc->m_currentPosition = targetPosition;
 	mc->m_currentVelocity = mc->m_velocity;
 }
 
-void StandardMovementMode::computeInterpolatedPosition(float alpha, GameObject* gameObject, MovementComponent* movementComponent) {
+void StandardMovementMode::computeInterpolatedPosition(double alpha, GameObject* gameObject, MovementComponent* movementComponent) {
 	GameObject* go = gameObject;
 	MovementComponent* mc = movementComponent;
-	go->m_position = (mc->m_currentPosition * alpha) + (mc->m_previousPosition * (1.0f - alpha));
+	go->m_position = (mc->m_currentPosition * alpha) + (mc->m_previousPosition * (1.0 - alpha));
 }
